refactor: 인증 관련 Configuration 통합
