--- src/main/java/woowacourse/auth/config/AuthenticationPrincipalConfig.java
@@ -1,22 +1,0 @@
-package woowacourse.auth.config;
-
-import java.util.List;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import woowacourse.auth.ui.AuthenticationPrincipalArgumentResolver;
-
-@Configuration
-public class AuthenticationPrincipalConfig implements WebMvcConfigurer {
-    private final AuthenticationPrincipalArgumentResolver authenticationPrincipalArgumentResolver;
-
-    public AuthenticationPrincipalConfig(
-            final AuthenticationPrincipalArgumentResolver authenticationPrincipalArgumentResolver) {
-        this.authenticationPrincipalArgumentResolver = authenticationPrincipalArgumentResolver;
-    }
-
-    @Override
-    public void addArgumentResolvers(final List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(authenticationPrincipalArgumentResolver);
-    }
-}
